rusk: Get rid of ServiceLoader
Move ProverService logic out of ServiceLoader to benefit of `Rusk` reference
